debugging

+++ inst/include/epiworld/tool-meat.hpp
@@ -490,10 +490,10 @@ inline void Tool<TSeq>::print() const
 
     printf_epiworld("Tool           : %s\n", tool_name->c_str());
     printf_epiworld("Id             : %s\n", (id < 0)? std::string("(empty)").c_str() : std::to_string(id).c_str());
-    printf_epiworld("state_init    : %i\n", state_init);
-    printf_epiworld("state_post    : %i\n", state_post);
-    printf_epiworld("queue_init     : %i\n", queue_init);
-    printf_epiworld("queue_post     : %i\n", queue_post);
+    printf_epiworld("state_init    : %i\n", static_cast<int>(state_init));
+    printf_epiworld("state_post    : %i\n", static_cast<int>(state_post));
+    printf_epiworld("queue_init     : %i\n", static_cast<int>(queue_init));
+    printf_epiworld("queue_post     : %i\n", static_cast<int>(queue_post));
 
 }
